Preserve comments when applying chart patches (#93)
Currently, we're losing comments in `values.yaml` everytime we apply a patch to it, since we have to roundtrip the YAML through JSON which drops all comments and also changes order of the fields.

This fixes that by using a library that applies the respective JSON patch to a YAML AST, preserving the order of fields and comments through the patching process.