Preserve metadata when patching Chart.yaml (#116)
The Chart.yaml contents are copied into the OCI artifact's
configuration. However, we're currently using the Chart.yaml contents
from *before we patch it*, meaning the data in the OCI config can be
inconsistent with the Chart.yaml in the OCI layer if a patch is applied. 

This re-orders the logic so that patched Chart.yaml data is used in
the to populate the OCI configuration.